JAMES-1862 Remove abstract level in StartTls tests
